simplifying minizip
