+++ cpp/open3d/data/extract/Extract.h
@@ -37,11 +37,5 @@ bool Extract(const std::string& filename,
              const bool always_overwrite = true,
              const bool print_progress = true);
 
-bool ExtractFromZIP(const std::string& filename,
-                    const std::string& extract_dir,
-                    const std::string& password = "",
-                    const bool always_overwrite = true,
-                    const bool print_progress = true);
-
 }  // namespace data
 }  // namespace open3d
